backend/requirements.txt
fuzzywuzzy
levenshtein

selenium~=4.8.3
beautifulsoup4~=4.12.0
undetected_chromedriver

Django~=4.1.3
requests~=2.28.1
djangorestframework~=3.14.0
drf_yasg~=1.21.5